requirements.txt
google-api-python-client==2.125.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
scikit-learn
nltk==3.8.1
pandas==2.2.2
google-generativeai==0.5.4
